feat : 믹스패널 이벤트 추가
