+++ usecase/src/main/java/org/depromeet/spot/usecase/service/review/CreateReviewService.java
@@ -13,6 +13,8 @@ import org.depromeet.spot.usecase.service.member.processor.MemberLevelProcessor;
 import org.depromeet.spot.usecase.service.review.processor.ReviewCreationProcessor;
 import org.depromeet.spot.usecase.service.review.processor.ReviewImageProcessor;
 import org.depromeet.spot.usecase.service.review.processor.ReviewKeywordProcessor;
+import org.depromeet.spot.usecase.service.util.MixpanelUtil;
+import org.depromeet.spot.usecase.service.util.MixpanelUtil.MixpanelEvent;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -30,6 +32,7 @@ public class CreateReviewService implements CreateReviewUsecase {
     private final ReviewImageProcessor reviewImageProcessor;
     private final ReviewKeywordProcessor reviewKeywordProcessor;
     private final MemberLevelProcessor memberLevelProcessor;
+    private final MixpanelUtil mixpanelUtil;
 
     @Override
     @Transactional
@@ -45,6 +48,9 @@ public class CreateReviewService implements CreateReviewUsecase {
         savedReview.setKeywordMap(keywordMap);
 
         Member levelUpdateMember = memberLevelProcessor.calculateAndUpdateMemberLevel(member);
+
+        // 믹스패널 이벤트(후기 등록 완료) 호출
+        mixpanelUtil.track(MixpanelEvent.REVIEW_REGISTER, String.valueOf(memberId));
 
         return new CreateReviewResult(savedReview, levelUpdateMember, review.getSeat());
     }
