+++ usecase/src/main/java/org/depromeet/spot/usecase/service/review/like/ReviewLikeService.java
@@ -6,6 +6,8 @@ import org.depromeet.spot.usecase.port.in.review.ReadReviewUsecase;
 import org.depromeet.spot.usecase.port.in.review.UpdateReviewUsecase;
 import org.depromeet.spot.usecase.port.in.review.like.ReviewLikeUsecase;
 import org.depromeet.spot.usecase.port.out.review.ReviewLikeRepository;
+import org.depromeet.spot.usecase.service.util.MixpanelUtil;
+import org.depromeet.spot.usecase.service.util.MixpanelUtil.MixpanelEvent;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,6 +21,7 @@ public class ReviewLikeService implements ReviewLikeUsecase {
     private final ReadReviewUsecase readReviewUsecase;
     private final UpdateReviewUsecase updateReviewUsecase;
     private final ReviewLikeRepository reviewLikeRepository;
+    private final MixpanelUtil mixpanelUtil;
 
     // TODO: 분산락 적용 예정
     @Override
@@ -31,6 +34,9 @@ public class ReviewLikeService implements ReviewLikeUsecase {
         }
 
         addLike(memberId, reviewId, review);
+
+        // 믹스패널 이벤트(좋아요 수) 발생
+        mixpanelUtil.track(MixpanelEvent.REVIEW_LIKE_COUNT, String.valueOf(memberId));
     }
 
     public void cancelLike(final long memberId, final long reviewId, Review review) {
