+++ usecase/src/main/java/org/depromeet/spot/usecase/service/util/MixpanelUtil.java
@@ -1,4 +1,4 @@
-package org.depromeet.spot.application.mixpanel;
+package org.depromeet.spot.usecase.service.util;
 
 import java.io.IOException;
 
@@ -10,6 +10,7 @@ import com.mixpanel.mixpanelapi.ClientDelivery;
 import com.mixpanel.mixpanelapi.MessageBuilder;
 import com.mixpanel.mixpanelapi.MixpanelAPI;
 
+import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
@@ -19,11 +20,14 @@ public class MixpanelUtil {
     @Value("${mixpanel.token}")
     private String mixpanelToken;
 
+    // mixpanelEvent는 eventName(이 단위로 이벤트가 묶임)
+    // distinctId는 사용자를 구분하는 데 사용됨.
     public void track(MixpanelEvent mixpanelEvent, String distinctId) {
         try {
 
             // 믹스패널 이벤트 메시지 생성
             MessageBuilder messageBuilder = new MessageBuilder(mixpanelToken);
+
             // 이벤트 생성
             JSONObject sentEvent = messageBuilder.event(distinctId, mixpanelEvent.getValue(), null);
 
@@ -34,9 +38,24 @@ public class MixpanelUtil {
             // Mixpanel로 데이터 전송
             MixpanelAPI mixpanel = new MixpanelAPI();
             mixpanel.deliver(delivery);
-            log.info("{} 정상 동작 완료", mixpanelEvent.getValue());
         } catch (IOException e) {
             e.printStackTrace();
+        }
+    }
+
+    @Getter
+    public enum MixpanelEvent {
+        REVIEW_REGISTER("review_register"),
+        REVIEW_REGISTER_MAX("review_register"),
+        REVIEW_OPEN_COUNT("review_open_count"),
+        REVIEW_LIKE_COUNT("review_like_count"),
+        REVIEW_SCRAP_COUNT("review_scrap_count"),
+        ;
+
+        String value;
+
+        MixpanelEvent(String value) {
+            this.value = value;
         }
     }
 }
